# web/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Understudy",
    "description": "Rehearse-then-dial voice recruiting on the Hunar Voice Agents API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List jobs",
        "operationId": "list_jobs_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobRead"
                  },
                  "type": "array",
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create a job",
        "description": "Stores the raw job description. Nothing is compiled yet \u2014 PUT /jobs/{id}/requirements does that.",
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get one job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/requirements": {
      "put": {
        "tags": [
          "jobs"
        ],
        "summary": "Update a job's requirements",
        "description": "Recompiles the raw JD and creates a new draft AgentVersion (origin=COMPILED, unpublished) for every language the compiled JD implies. Never edits an existing version \u2014 versions are immutable (CLAUDE.md).",
        "operationId": "update_requirements_jobs__job_id__requirements_put",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementsUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/versions": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Version history with composite score per version",
        "operationId": "list_versions_jobs__job_id__versions_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionHistoryRow"
                  },
                  "title": "Response List Versions Jobs  Job Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/versions/{version_no}/publish": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Publish a draft version to Hunar",
        "description": "`language` disambiguates: versions are numbered per (job, language), so version_no alone is not unique across languages.",
        "operationId": "publish_job_version_jobs__job_id__versions__version_no__publish_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "version_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version No"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/personas": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "The six rehearsal personas",
        "operationId": "list_personas_jobs__job_id__personas_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PersonaRead"
                  },
                  "title": "Response List Personas Jobs  Job Id  Personas Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/source": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Source candidates",
        "description": "Every field falls back to the job's own compiled search_query when omitted, so re-sourcing more of the same needs no body at all.",
        "operationId": "source_candidates_jobs__job_id__source_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/candidates": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List candidates, best match first",
        "operationId": "list_candidates_jobs__job_id__candidates_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateRead"
                  },
                  "title": "Response List Candidates Jobs  Job Id  Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/call": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Launch outbound screening calls",
        "description": "Blocked candidates are returned WITH reasons, never silently skipped \u2014 the consent/DNC guard is unbypassable (CLAUDE.md).",
        "operationId": "launch_calls_jobs__job_id__call_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallLaunchSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/board": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "The live candidate/call board",
        "description": "Refreshes non-terminal outreach rows on read (polling), then reads whatever state is current \u2014 the board is correct even if Hunar's webhooks never arrive.",
        "operationId": "get_board_jobs__job_id__board_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/export": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Export candidates as CSV",
        "description": "One row per candidate, one column per screening question id plus the four standard result fields, pulled from each candidate's latest call result.",
        "operationId": "export_candidates_jobs__job_id__export_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreate": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title"
          },
          "raw_jd": {
            "type": "string",
            "minLength": 1,
            "title": "Raw Jd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "raw_jd"
        ],
        "title": "JobCreate"
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "raw_jd": {
            "type": "string",
            "title": "Raw Jd"
          },
          "compiled": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compiled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "raw_jd",
          "compiled",
          "created_at"
        ],
        "title": "JobRead"
      },
      "Language": {
        "type": "string",
        "enum": [
          "ENGLISH",
          "HINDI",
          "TAMIL",
          "TELUGU",
          "KANNADA",
          "MARATHI",
          "MALAYALAM",
          "GUJARATI",
          "BENGALI",
          "TURKISH",
          "ARABIC",
          "SPANISH"
        ],
        "title": "Language",
        "description": "Belongs to the agent, not the call \u2014 see CLAUDE.md."
      },
...
      "RequirementsUpdate": {
        "properties": {
          "raw_jd": {
            "type": "string",
            "minLength": 1,
            "title": "Raw Jd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "raw_jd"
        ],
        "title": "RequirementsUpdate",
        "description": "Body for PUT /jobs/{id}/requirements \u2014 a revised raw JD. Recompiling always creates new\ndraft AgentVersion row(s); it never edits an existing one (CLAUDE.md: versions immutable)."
      },
      "RequirementsUpdateResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "job_id",
          "versions"
        ],
        "title": "RequirementsUpdateResponse"
      },
...
      "SourceResponse": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateRead"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "cached",
          "candidates"
        ],
        "title": "SourceResponse"
      },
...
    }
  },
...
}